Follow Java coding standard
Modified class/enum names and variables to match naming conventions.
Added all project files to their relavant packages
like `task` or `exception`.
Added relavant class header comments and Javadocs for understandability.

Additionally, improved readability by abstracting setCommand function.
