+++ src/main/java/ip/exception/InvalidCommand.java
@@ -1,0 +1,5 @@
+package ip.exception;
+
+public class InvalidCommand extends RobotException {
+    
+}
