+++ src/main/java/ip/exception/NoDeadline.java
@@ -1,3 +1,5 @@
+package ip.exception;
+
 public class NoDeadline extends RobotException {
     
 }
