+++ src/main/java/ip/exception/NoDescription.java
@@ -1,3 +1,5 @@
+package ip.exception;
+
 public class NoDescription extends RobotException {
     
 }
