+++ src/main/java/ip/exception/NoPeriod.java
@@ -1,3 +1,5 @@
+package ip.exception;
+
 public class NoPeriod extends RobotException {
     
 }
