+++ src/main/java/ip/exception/NoTaskFound.java
@@ -1,3 +1,5 @@
+package ip.exception;
+
 public class NoTaskFound extends RobotException {
     
 }
