+++ src/main/java/ip/exception/RobotException.java
@@ -1,3 +1,5 @@
+package ip.exception;
+
 public class RobotException extends Exception {
 
     public RobotException() {
