+++ src/main/java/ip/task/ToDo.java
@@ -1,4 +1,8 @@
+package ip.task;
+
 import java.util.Scanner;
+
+import ip.exception.NoDescription;
 
 public class ToDo extends Task {
